remove Cppcheck XML format V1 support - V1 report support is deprecated: use XML V2 format instead (--xml-version=2) - support Cppcheck V1.69++ (remove messages from former versions) - close #1247
