--- cxx-sensors/src/test/resources/org/sonar/cxx/sensors/reports-project/cppcheck-reports/cppcheck-result-badlines-V1.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0"?>
-<results>
-  <error file="sources/utils/code_chunks.cpp" line="boom" id="unusedFunction" severity="style" msg="The function 'foo' is never used"/>
-  <error file="sources/utils/code_chunks.cpp" line="-1" id="unusedFunction" severity="style" msg="The function 'foo' is never used"/>
-</results>
