--- cxx-sensors/src/test/resources/org/sonar/cxx/sensors/reports-project/cppcheck-reports/cppcheck-result-projectlevelviolation-V1.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0"?>
-<results>
-<error id="id1" severity="error" msg="somemessage"/>
-<error id="id2" file="*" severity="error" msg="somemessage"/>
-<error id="id3" file="*" line="0" severity="error" msg="somemessage"/>
-</results>
