+++ cxx-sensors/src/tools/generate_cppcheck_resources.cmd
@@ -4,8 +4,7 @@ SET SCRIPT_DIR=%~dp0
 SET CPPCHECK_DIR=C:\Program Files\Cppcheck
 SET PYTHON_DIR=C:\Perforce\Helix-QAC-2019.2\components\python-2.7.10
 
-SET CPPCHECK_LIBRARY_ARGS=--library=avr.cfg --library=boost.cfg --library=bsd.cfg --library=cppunit.cfg --library=embedded_sql.cfg --library=gnu.cfg --library=googletest.cfg --library=gtk.cfg --library=libcerror.cfg --library=microsoft_sal.cfg --library=motif.cfg --library=nspr.cfg --library=opengl.cfg --library=posix.cfg --library=python.cfg --library=qt.cfg --library=ruby.cfg --library=sdl.cfg
---library=sfml.cfg --library=sqlite3.cfg --library=std.cfg --library=windows.cfg --library=wxwidgets.cfg --library=zlib.cfg
+SET CPPCHECK_LIBRARY_ARGS=--library=avr.cfg --library=boost.cfg --library=bsd.cfg --library=cppunit.cfg --library=embedded_sql.cfg --library=gnu.cfg --library=googletest.cfg --library=gtk.cfg --library=libcerror.cfg --library=microsoft_sal.cfg --library=motif.cfg --library=nspr.cfg --library=opengl.cfg --library=posix.cfg --library=python.cfg --library=qt.cfg --library=ruby.cfg --library=sdl.cfg --library=sfml.cfg --library=sqlite3.cfg --library=std.cfg --library=windows.cfg --library=wxwidgets.cfg --library=zlib.cfg
 
 rem wget https://cwe.mitre.org/data/xml/cwec_v3.2.xml.zip --output-document=cwec_v3.2.xml.zip && unzip -j -o cwec_v3.2.xml.zip
