+++ integration-tests/testdata/cppcheck_project/relative-to-src.xml
@@ -1,9 +1,24 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<results>
-<error file="component1.cc" line="15" id="unusedVariable" severity="style" msg="Unused variable: x"/>
-<error file="component1.cc" line="35" id="unreadVariable" severity="style" msg="Variable 'i' is assigned a value that is never used."/>
-<error file="component1.cc" line="45" id="deallocDealloc" severity="error" msg="Deallocating a deallocated pointer: ip"/>
-<error file="component1.cc" line="45" id="doubleFree" severity="error" msg="Memory pointed to by 'ip' is freed twice."/>
-<error file="component1.cc" line="30" id="uninitvar" severity="error" msg="Uninitialized variable: a"/>
-<error file="component1.cc" line="22" id="unusedFunction" severity="style" msg="The function 'do_valgrind_errors' is never used."/>
+<results version="2">
+  <cppcheck version="1.61"/>
+  <errors>
+    <error  id="unusedVariable" severity="style" msg="Unused variable: x">
+      <location file="component1.cc" line="15"/>
+    </error>
+    <error  id="unreadVariable" severity="style" msg="Variable 'i' is assigned a value that is never used.">
+      <location file="component1.cc" line="35"/>
+    </error>
+    <error  id="deallocDealloc" severity="error" msg="Deallocating a deallocated pointer: ip">
+      <location file="component1.cc" line="45"/>
+    </error>
+    <error  id="doubleFree" severity="error" msg="Memory pointed to by 'ip' is freed twice.">
+      <location file="component1.cc" line="45"/>
+    </error>
+    <error  id="uninitvar" severity="error" msg="Uninitialized variable: a">
+      <location file="component1.cc" line="30"/>
+    </error>
+    <error id="unusedFunction" severity="style" msg="The function 'do_valgrind_errors' is never used.">
+      <location file="component1.cc" line="22"  />
+    </error>
+  </errors>
 </results>
