+++ integration-tests/testdata/cppcheck_project/sonar-project.properties
@@ -16,4 +16,4 @@ sonar.sources=src
 sonar.cxx.includeDirectories=src
 
 # paths to the reports
-sonar.cxx.cppcheck.reportPath=cppcheck-v1.xml
+sonar.cxx.cppcheck.reportPath=cppcheck-v2.xml
