+++ integration-tests/testdata/cppcheck_project_c_cpp/sonar-project.properties
@@ -12,4 +12,4 @@ sonar.sources=src
 sonar.cxx.includeDirectories=src
 
 # paths to the reports
-sonar.cxx.cppcheck.reportPath=cppcheck-v1.xml
+sonar.cxx.cppcheck.reportPath=cppcheck-v2.xml
