+++ integration-tests/testdata/googletest_bullseye_vs_project/sonarcpp/reports-cppcheck/cppcheck-result-0.xml
@@ -1,6 +1,16 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<results>
-    <error file="./PathHandling\PathHandle.cpp" line="15" id="unreadVariable" severity="style" msg="Variable &apos;pFile&apos; is assigned a value that is never used."/>
-    <error file="./PathHandling\PathHandle.cpp" line="17" id="unreadVariable" severity="style" msg="Variable &apos;i&apos; is assigned a value that is never used."/>
-    <error id="missingInclude" severity="style" msg="Cppcheck cannot find all the include files (use --check-config for details)"/>
+<results version="2">
+  <cppcheck version="1.61"/>
+  <errors>
+    <error id="unreadVariable" severity="style" msg="Variable &apos;pFile&apos; is assigned a value that is never used.">
+      <location file="./PathHandling\PathHandle.cpp" line="15" />
+    </error>
+    <error id="unreadVariable" severity="style" msg="Variable &apos;i&apos; is assigned a value that is never used.">
+      <location file="./PathHandling\PathHandle.cpp" line="17"/>
+    </error>
+    <error id="missingInclude" severity="style" msg="Cppcheck cannot find all the include files (use --check-config for details)">
+      <location file="*" />
+    </error>
+  </errors>
 </results>
+
